fix: dev-container setup (#1771)
* fix: webpacker container startup

Enable clean webpacker startup by ensuring intended startup order of containers.

* fix: devcontainer

With https://github.com/ComPlat/chemotion_ELN/pull/1665 the node installation is moved from `Dockerfile.chemotion-dev` to `run-ruby-dev.sh`, which is not executed by the devcontainer. That's why the devcontainer would not start. Running `run-ruby-dev.sh` as part of the devcontainer startup would introduce a lifetime depdendency from the rails server (running in the `app` container) to the devcontainer. This is why we extracted the installation of node into a new script, which is now called from `run-ruby-dev.sh` as well as `postCreateCommand` command in devcontainer.

* Don't start `webpacker` before all gems are installed in `app`

---------

Co-authored-by: Sören Peters <dev.peters@icloud.com>
Co-authored-by: Jan C. Brammer <jan.c.brammer@gmail.com>